backend/requirements.txt
# ---------------------------------------------------------------------------
# Web Framework
# ---------------------------------------------------------------------------
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
python-multipart==0.0.9
httpx==0.27.0

# ---------------------------------------------------------------------------
# Database (MySQL)
# ---------------------------------------------------------------------------
sqlalchemy==2.0.30
alembic==1.13.1
aiomysql==0.2.0
PyMySQL==1.1.1
cryptography==42.0.7

# ---------------------------------------------------------------------------
# Redis / Caching
# ---------------------------------------------------------------------------
redis==5.0.4
aioredis==2.0.1

# ---------------------------------------------------------------------------
# Task Queue
# ---------------------------------------------------------------------------
celery==5.4.0
flower==2.0.1

# ---------------------------------------------------------------------------
# Authentication
# ---------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ---------------------------------------------------------------------------
# Validation / Settings
# ---------------------------------------------------------------------------
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
python-dotenv==1.0.1

# ---------------------------------------------------------------------------
# ML / Deep Learning
# ---------------------------------------------------------------------------
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2

# ---------------------------------------------------------------------------
# Image Processing
# ---------------------------------------------------------------------------
opencv-python-headless==4.9.0.80
Pillow==10.3.0

# ---------------------------------------------------------------------------
# NLP / Text
# ---------------------------------------------------------------------------
nltk==3.8.1

# ---------------------------------------------------------------------------
# PDF Processing
# ---------------------------------------------------------------------------
PyMuPDF==1.24.3
reportlab==4.2.0
python-docx==1.1.2

# ---------------------------------------------------------------------------
# Utilities
# ---------------------------------------------------------------------------
python-slugify==8.0.4
structlog==24.1.0
tenacity==8.3.0
rich==13.7.1
tqdm==4.66.4
pytz==2024.1
orjson==3.10.3

# ---------------------------------------------------------------------------
# Testing
# ---------------------------------------------------------------------------
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
factory-boy==3.3.0
Faker==25.2.0

# ---------------------------------------------------------------------------
# Dev / Linting
# ---------------------------------------------------------------------------
ruff==0.4.4
black==24.4.2
mypy==1.10.0
pre-commit==3.7.1